Add commands concept to wireit test rig (#3)
The wireit test rig can now generate commands, which are temporary unique scripts created for the purpose of testing, taking the place of what would be binaries like tsc, rollup, etc. in an actual build config.

When a command is invoked, it makes an IPC connection back to the rig, so that we can keep track of when it's been spawned, and tell it when to exit.

This will let us run end-to-end tests to ensure that wireit is invoking scripts in the way we expect.