Add events, errors, and loggers (#19)
Event
----------
Adds an Event type, which contains structured information about events that can happen during the course of Wireit's execution. This includes Success, Failure, Output, and Info events. The purpose is to separate semantic logging from presentation, so that we can implement different approaches for handling errors, script output, and other info that the user may care about.

For example:

{
  type: 'success',
  reason: 'exit-zero',
  script: {
    package: '/some/package/dir',
    script: 'build',
  }
}

{
  type: 'failure',
  reason: 'exit-non-zero',
  status: 42,
  script: {
    package: '/some/package/dir',
    script: 'build',
  }
}

{
  type: 'output',
  stream: 'stdout',
  data: 'some output from the build script',
  script: {
    package: '/some/package/dir',
    script: 'build',
  }
}

Logger
----------
Adds a Logger interface, and a DefaultLogger class. Loggers receive events and display events to the user in some way. The DefaultLogger contains custom formatted messages for each kind of Event and writes them to stdout/stderr. Other Loggers we could add in the future could be things like a tmux one, one that logs to files, etc.

WireitError
----------
The WireitError class is a subclass of Error which has an event property which must be a Failure event. If a WireitError is thrown during execution, then its event will ultimately get passed to the logger. This lets us use exceptions for control flow, but still standardize on the Event type for representing all errors.