Self host with a known good tarball (#53)
Wireit is now good enough to build and test itself!

Last good tarball
=================

However, we don't want to use the very latest version of the code during development, because if we break something, we'll be unable to build or test, or might build or test incorrectly.

Instead, this PR adds a new `known-good.tgz` tarball, created by running `npm pack`. It's included as a `devDependency`, so the known good version is installed to `node_modules/`, which is independent from the code in `lib/` etc. When we invoke `wireit` from our `scripts`, it refers to the one in `node_modules/`, not `bin/`.

For now, we'll update this tarball manually with `npm run update-known-good`. In the future, we'll probably want to bump this known-good tarball after each release, possibly automatically.

Using Wireit
============

We now use Wireit to `build` and `test` itself. E.g., to automatically re-build and re-test on every change, you can now do:

  npm test watch

This also lets us remove `npm run build` from the `tests.yml` file -- since `npm test` now automatically builds.