Add replays of stdout and stderr (#68)
When a script is fresh, we now replay any `stdout` or `stderr` that it produced in the previous run. Previously when a script was fresh, nothing would be emitted (apart from the log events from Wireit itself).

The replay files are written to `<package-dir>/.wireit/<script-name-hex>/(stdout|stderr)`.

This is critical for use cases where the output of a command is being consumed downstream, and it's also helpful for humans. For example, if some tests previously passed, you will now see the test runner output, instead of nothing.

Fixes https://github.com/lit/wireit/issues/61

Also
====

- Factored out a `ScriptExecution` class from `Executor`. Almost every method was receiving a `script` parameter, so this makes things much more concise.