Update PATH depending on current package (#85)
Before this PR, we never modified the `PATH` environment variable of the Wireit process, so the shell processes we spawned were only be able to find binaries in the `node_modules/.bin` directory of the package we *started* in (and its ancestors).

This is fine if a dependency is in the same package, but not if it's in another package, because that will cause binaries that are installed only to the dependency's `node_modules/` folder to not be found.

We now generate a correct `PATH` variable as part of spawning a script's command, which includes the binaries that are accessible from that package, and excludes the ones that are only accessible from the starting package. This simulates running `npm run <script>` from each package as we cross package boundaries.

Fixes https://github.com/google/wireit/issues/81

Also
====

Disabled `fail-fast` on CI. The default GitHub Actions behavior is that as soon as one matrix configuration fails, all other ones are cancelled. This can make it harder to understand which OS's are affected by a failure, and in which way.