Improve logo in dark mode (#59)
The "Wireit" part of the logo now has a white outline, so that on dark backgrounds it is visible. It looks the same on a white background.

Also slightly changed the color of the squiggle for better contrast on dark backgrounds.

Also smoothed out the W, and evened out the baselines of some of the other letters.