Make order of glob !exclusions significant (#109)
fast-glob doesn't pay attention to the order of `!excluded` patterns. For example, the following pattern array should include `foo/bar/baz`, but fast-glob excludes it:

```
foo/**
!foo/bar/**
foo/bar/baz  <-- wrongly excluded
!foo/qux
```

That's because fast-glob simply takes all `!exclusion` patterns, without paying attention to where they appear in the array. We actually want to support *re-inclusion*, where a file is excluded by one pattern, and then re-included with a subsequent pattern. This is a useful feature, and it's the behavior of the `files` array in `package.json`, so I think it's how we should work too.

To fix this behavior, we divide the patterns into groups that can be evaluated separately, and then combined at the end. We create a new group whenever an `!exclude` pattern is in front of an include pattern, because that's when this problem could occur, and we include all subsequent exclusions (but not preceding ones) in each group:

```
Group 1:
  include: foo/**
  exclude: foo/bar/**
  exclude: foo/qux

Group 2:
  include: foo/bar/baz
  exclude: foo/qux
```

Fixes https://github.com/google/wireit/issues/108