Return glob results as objects instead of strings (#116)
`fast-glob` can return objects instead of strings. These objects contain the `path` along witha a [`dirent`](https://nodejs.org/api/fs.html#class-fsdirent) which can be used to check if something is a regular file, a directory, or some other kind of file.

According to https://github.com/mrmlnc/fast-glob#objectmode there is no overhead for returning this object, because it is an internal representation that `fast-glob` already uses (note that [`readdir`](https://nodejs.org/api/fs.html#fspromisesreaddirpath-options) can return dirents).

We now enable this option and return these objects from our `glob` function.

The context for this is upcoming PRs to address https://github.com/google/wireit/issues/77, which will switch to a different approach for deleting and copying files. This new approach will need to know whether each file is a regular file versus a directory, and it will work faster if it already receives that information, rather than having to `stat` for it.