Only delete directories if they are empty (#117)
Previously, when cleaning output, we used `fs.rm` to recursively delete all matching files and directories matched by the `output` glob patterns.

The problem with that approach was that if a directory was matched in the glob patterns, but contained a file that was supposed to be `!excluded`, we would wrongly greedily delete the excluded file when we recursively deleted the directory.

Now, we instead expand all directories explicitly while globbing, and replace `fs.rm` with our own implementation. The new implementation only deletes directories if they have no children after all of the other listed child files/directories have been deleted.

Partly fixes https://github.com/google/wireit/issues/77 (the same issue still exists for caching, which will be addressed in the next PR).