Fix package location bug when script is in npm workspace (#124)
Previously, we were using the `npm_config_local_prefix` environment variable to discover the package directory for the current script.

However, when a package is a member of a workspace (meaning a parent directory has a `package.json` that lists the path of that package in its `workspaces` array), then `npm_config_local_prefix` instead corresponds to the workspace root package.

We now instead use the `npm_package_json` environment variable, which in the case of a workspace is set to the specific workspace package, instead of the root. The only other difference is that it includes the `/package.json` portion of the path.

(It's interesting to note that this bug affected not only commands like `npm run foo -ws`, but a regular command like `cd packages/foo; npm run foo`. This implies that npm actually walks up the filesystem tree to discover if there is another `package.json` file that lists the package as one of its workspaces.)

Fixes https://github.com/google/wireit/issues/123