Solidify symlink behavior (#129)
The symlink behavior was unclear previously. Now it is explicit and better tested. The behavior is that we follow symlinks for input `files`, but not for `output` files.

The reason we follow symlinks for input `files` is that the content at the target of the symlink is likely to affect the behavior of the script.

The reason we don't follow symlinks for `output` files is that when restoring from cache, we want to restore the exact filesystem output created by the script. If a script created a symlink, then we want to restore a symlink -- not a copy of the file/directory the symlink pointed to.

One interesting thing I learned is that Windows symlinks need to be explicitly flagged if they are links to directories. You can't walk through them otherwise. https://docs.microsoft.com/en-us/windows-server/administration/windows-commands/mklink (see `/d` flag).

Fixes https://github.com/google/wireit/issues/54