Don't cache excluded files (#134)
Previously, if a directory was included in the `output` config, then we would cache all of that directory's children, even if there were `!exclude` patterns that should have excluded some or all of its children.

For example, in this config, we would wrongly cache the `output/excluded` file:

```json
{
  "output": [
    "output",
    "!output/excluded"
  ]
}
```

The reason this happened was very similar to the reason behind https://github.com/google/wireit/pull/117. We were copying using recursive operations (`fs.cp` for local cache, and `tar` for GitHub cache), so when a directory was listed, the exclusion patterns ended up making no difference.

Now, we instead explicitly expand all directories when we generate the list of files to cache, and pass them to our own custom `copyEntries` function, which does NOT copy recursively. Empty directories work too, because directories are explicitly included in the glob results, and we `mkdir` them directly.

The one caveat is that this PR does not support empty directories on GitHub Actions, because we can't control the fact that `tar` includes directories recursively. The fix for that is in https://github.com/google/wireit/pull/135 (seemed like it deserved its own PR).

Penultimate PR for https://github.com/google/wireit/issues/77

### Also

- Adds some type-safety around relative vs absolute results being returned from `glob`.

- A side effect of this change is that our local cache now caches symlink targets verbatim. The built-in `fs.cp` function resolves symlink targets to absolute paths. I think verbatim is the better behavior, because you want to match the output as closely as possible, and that also matches what the GitHub cache implementation already did.