Support caching empty directories in GitHub Actions cache (#135)
Follow up to https://github.com/google/wireit/pull/134

When we cache with GitHub Actions, we build a tarball. We do this by passing paths to the `tar` command. However, when we pass a *directory* to `tar`, all of its children are implicitly included. This is a problem, because sometimes we want to add a directory, but only a subset of its children (or none of them), because of exclusion patterns in the `output` globs.

To work around this problem, we never pass directories to `tar`. Instead, we enumerate every specific file. This works fine, **except for empty directories**. If a directory was explicitly listed for inclusion, and it happens to be empty, we should cache an empty directory.

So we need special handling for empty directories. If there are empty directories, we create a special "empty directories manifest" file, and include that in the tarball at a predictable location (inside the `.wireit` folder so that it won't collide with a user file). Then when we untar, we check for that manifest, create any missing directories listed inside it, and delete the manifest.

Fixes https://github.com/google/wireit/issues/77