Upgrade our own version of wireit and enable GitHub caching (#141)
Bumps the version of Wireit we use to test Wireit to 0.2.0, and enables GitHub Actions caching.

The 2nd commit in this PR is a minor README fix. You can see that it finished in only a few seconds, because all of the tests were already cached from the first commit.