Build extension into subdirectory
This resolves an issue with the latest version of npm and npm workspaces and vsce, where vsce thinks that the root of the extension is the root of the workspace and adds too many files to the extension.

It also forces us to be very deliberate about what files are included in the extension.

Finally, it lets us separate out the package.json file that we use for scripts and devdependencies from the file we use to configure the extension, which means that we can give the extension an id that matches its github repo: google.wireit