Unset github caching environment variables in tests (#159)
Fixes an oversight from when I turned on GitHub Actions caching in https://github.com/google/wireit/pull/141.

The `GITHUB_CACHE` environment variable that was turned on propagated to all of the Wireit instances that we spawned in tests.

That meant we were actually using live, remote caching for every test that had `files` and `output` defined. Oops!