Add an offset to position converter
Parsers, including jsonc-parser, often give locations in terms of a character offset, but when communicating to users we want to use line and character numbers.

This also introduces a JsonFile object, which includes the path, the string contents, and the ast of a json file all in one object. Useful for passing around, and also, in an upcoming change, as a WeakMap key to cache the OffsetToPosition converter from.