Interpret leading slashes relative to package (#180)
Previously, a glob pattern like `/foo` was resolved to `foo` in the root of the filesystem, which is the usual behavior for a file path.

However, that's a footgun because what a user almost certainly means by that is `foo` in the package directory.

So we now ignore the leading `/`. In the case of `files`, we still allow reaching out of the directory using e.g. `../foo`.

This new behavior also matches the behavior of the npm `files` array (apart from the `../` support, which `files` simply ignores).

Also fixed a duplicate `lint` script in `package.json`.

Fixes https://github.com/google/wireit/issues/173