Serialize execution of the same script across different Wireit processes (#189)
Previously, if you `npm run` multiple versions of the same script (or 2 Wireit scripts that share a dependency) at the same time, then they would simply run simultaneously, and incorrect output could be produced. In particular, the second Wireit process might try and delete output, while the first one is still writing it.

Now, we prevent multiple versions of the same script from running at the same time. Before any script starts, we try to acquire an exclusive lock for it, and we wait if the lock can't be acquired, logging this message:

```
💤  [a] Waiting for another process which is already running this script.
```

An exception is when `output=[]`, because I think these cases are unlikely to cause problems, and it allows users to continue running e.g. multiple versions of the same server or other non-build step if they need to.

The lockfile is managed using the [proper-lockfile](https://github.com/moxystudio/node-proper-lockfile) library, which seems to have a nice design. It works by attempting to `mkdir` a directory, which will either succeed if it doesn't exist, or fail it does exist, atomically. It also detects when a lock has probably been abandoned, by regularly updating the directory `mtime` when the lock is held, and deleting the lock directory if it hasn't been updated in a long time.

Fixes https://github.com/google/wireit/issues/35