Don't start new scripts after failure by default, add option for old behavior (#209)
### Before

Previously, if a script failed, a new script could still start as long as the failure didn't occur in its transitive dependencies.

For example in this graph, after `fail` fails, `bar` would not start, but `foo` would:

```
 main
  / \
 |   |
 v   v
foo  bar
      |
      v
     fail
```

### After

Now, by default, as soon as a failure occurs anywhere in the build graph, new scripts will not be started, but ones that are already running will be allowed to finish.

The old behavior is still useful, e.g. when you want to have a complete picture of which scripts are passing vs failing, so the old behavior can be enabled by setting:

```bash
WIREIT_FAILURES=continue
```

The default can be set explicitly with (not sure about this name, open to suggestions):

```bash
WIREIT_FAILURES=no-new
```

Fixes https://github.com/google/wireit/issues/29

### Follow up

In a follow-up PR, I am planning to add a `WIREIT_FAILURES=kill` mode, which will be even more aggressive, and terminate running scripts as soon as the first one fails. I think this is also useful in some situations, for when you want to be notified about a failure as soon as possible.