Add README content to website (#216)
This is not the final content that we should use even before publicizing the site, but it's a good starting point.

Also makes the website servable from any directory, instead of having to be built knowing what the directory is up-front, by using relative paths. Needed a custom filter for this, surprisingly. This lets us serve the same exact build from any of these mount points:

```
http://localhost:8080/
https://google.github.io/wireit/
https://google.github.io/wireit/previews/<pr>/
```