Active script termination (#214)
Previously, we depended on the fact that we didn't use [`detached`](https://nodejs.org/api/child_process.html#optionsdetached) when spawning scripts, so that we didn't need to do anything special when the user sent a `SIGINT` to Wireit (e.g. via Ctrl-C). However, this meant that we didn't have the ability to terminate just one script; either Wireit and all child processes were killed, or none where.

Now, we enable `detached` mode (except on Windows where it isn't needed and is actually bad), so that we can control termination of specific scripts ourselves, and add a `terminate` method to `ScriptChildProcess`.  This is a prerequisite for the next PR, which will add `WIREIT_FAILURES=kill` (https://github.com/google/wireit/issues/31).

Also includes refactoring to the `Executor` to more clearly handle the 3 related events: `failureOccured`, `stopStartingNewScripts`, and `killRunningScripts`.

Fixes https://github.com/google/wireit/issues/30