[ide] Be slightly more generous with certain locations
When we're looking for the meaning of a location in a file, instead of only counting the name and the value of a property, also count the colon and whitespace in between.

i.e. if you're trying to jump to the definition of a wireit config from the scripts section, currently you can do so from either:

```
    "a": "wireit"
    ~~~
```

or

```
    "a": "wireit"
         ~~~~~~~~
```

but after this change you can do it anywhere in:

```
    "a": "wireit"
    ~~~~~~~~~~~~~
```