Rename cachekey/state to fingerprint (#244)
When a script executes, a JSON object is generated for it which represents all of the inputs to that script at that time: its configuration, its input files and their content hashes, plus the inherited state of its dependencies. This value is then used to determine whether a script can be skipped, and as the key for cache entries.

Previously, this was variously referred to as either the "cache key" or the "state" of the script, and was saved to a file at `.wireit/<script>/state`.

Now, this concept is consistently referred to as the "fingerprint", and is saved to `.wireit/<script>/fingerprint`.

I think having a distinct consistent name will help readability and documentation. I noticed this term being used in the Turborepo docs for its equivalent concept, and it seems like a good choice.