Remove stdout/stderr replay (#310)
I have come to feel that the stdout/stderr replay is an anti-feature that adds a lot of noise and very little value.

I think there are some use cases where you might want it, e.g. where the whole purpose of running the script is to see some output on your terminal. But those are fairly rare. Perhaps we could add a per-script setting for those cases.

I think the right default is to be quieter, though, so I'm inclined to just remove the feature for now, and think some more about adding it back in a more thoughtful way in the future.