Handle 503 errors from GitHub the same as 429 errors (#326)
Sometimes we get 503 "Service Unavailable" errors from the GitHub Caching service that are out of our control. Previously, this would be fatal. Now, it is non-fatal, and instead just causes us to stop using the caching service for the remainder of that Wireit process. This is the same way we already handled 429 "Too Many Requests" errors.

Note in the future we could add some retries, but since caching is optional, this is a good enough behavior in the meantime.

Fixes https://github.com/google/wireit/issues/313