Add single task debounce to watcher (#437)
File change events are now debounced using a `setTimeout` with 0ms. In the future we could maybe make this a user-controlled parameter, but the immediate purpose is to solve a bug I noticed while setting up the webcomponents.org dev configuration.

The bug was that if multiple scripts have the same input files, then because in watch mode we register a filesystem watcher for *each of those scripts* *, when a common file changed, we'd start a run from the first event, and then immediately queue up a second run from the second event. Now we wait a task before running, so the two events get batched.

For a build with good caching configuration, this would usually cause a no-op second build -- but that's quite annoying because of the console spam. For a build without caching, this would cause a fully redundant build.

**(We could potentially register fewer watchers by noticing when input files are the same across scripts, but note that different globs can match the same file, so it's often not possible to do this, and the bug would still exist.)*