Shut down services that have been deleted between watch iterations (#489)
Handles the case where we're in watch mode, and a service was entirely deleted from the graph. Previously it would keep running, but now we notice it and shut it down before starting the next execution. The same goes for scripts that used to be "directly invoked" and are now no longer (since those might not need to run at all).

Part of https://github.com/google/wireit/issues/33