Allow service consumers to cache (#490)
Allows scripts that consume services to cache. This just requires slightly relaxing how we define when a service is "fully tracked". A standard script needs both its inputs and outputs known in order to be "fully tracked", but services never have output so we need to not require it in their case.