Rename triggersRerun to cascade, and add new documentation (#520)
First it was `soft`, then it was `triggersRerun`, now it's `cascade` (note we haven't released yet, so none of these changes were breaking).

This name came to me last night, and I kind of like it. I think `soft` is too vague and non-technical, but `triggersRerun` is just a mouthful (I've been typing it out a bunch, and it's been getting on my nerves). `cascade` feels like it sort of gives an impression of what the thing does, and I think will be memorable and understandable once you've read the docs.

Speaking of, I also rewrote the docs in a way that I hope is more clear.

Also adds some missing docs about `dependencies` and `service` to the configuration reference table.

Addon to https://github.com/google/wireit/issues/237